Extract HiveReaderUtil with Hive dependencies
