+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/GenericHiveRecordCursor.java
@@ -63,8 +63,8 @@ import static io.trino.plugin.base.util.Closables.closeAllSuppress;
 import static io.trino.plugin.hive.HiveColumnHandle.ColumnType.REGULAR;
 import static io.trino.plugin.hive.HiveErrorCode.HIVE_BAD_DATA;
 import static io.trino.plugin.hive.HiveErrorCode.HIVE_CURSOR_ERROR;
-import static io.trino.plugin.hive.util.HiveUtil.getDeserializer;
-import static io.trino.plugin.hive.util.HiveUtil.getTableObjectInspector;
+import static io.trino.plugin.hive.util.HiveReaderUtil.getDeserializer;
+import static io.trino.plugin.hive.util.HiveReaderUtil.getTableObjectInspector;
 import static io.trino.plugin.hive.util.HiveUtil.isStructuralType;
 import static io.trino.plugin.hive.util.SerDeUtils.getBlockObject;
 import static io.trino.spi.type.BigintType.BIGINT;
